readahead: common code for the implementation
Readahead is only used when the file access has been sequential.  This
is tracked in the POSIX-like API file handles.  The FSE API does not
support readahead at this time.

Additionally, the core determines whether readahead makes sense for a
given sequential read call: see Readahead() in inodedata.c.

For simplicity, when readahead is performed, the amount requested is
always a single block of data.  It might be worthwhile to make this a
configurable value.

For context, see issue #16 in the Reliance Edge repository on GitHub.