posix: don't register tasks in red_uninit()
Previously, if red_uninit() was called by a task which was not a file
system user, that task would be automatically registered as a file
system user as part of PosixEnter().  If all the task slots were already
in use, this registration would fail with a RED_EUSERS error, which is
problematic for two reasons: firstly, because it was undocumented
behavior, and secondly, because it could prevent red_uninit() from ever
succeeding, since there is no way to deregister the tasks whose
occupancy of the task slots prevents the current task from completing
the red_uninit().  Even if the RED_EUSERS error does not occur, the task
registration was pointless, since a successful uninitialization means
that all task registrations are null and void.

To avoid the possibility of a RED_EUSERS error, don't call PosixEnter(),
and hence don't register the calling task as a file system user.  We
still need to acquire the FS mutex, but that can be done by calling
RedOsMutexAcquire() directly.  Doing so is more symmetrical, since
red_uninit() was already calling RedOsMutexRelease() directly rather
than using PosixLeave().

Both before this change and afterward, the only case where red_uninit()
can register a task successfully is when it fails with a RED_EBUSY
error.  Previously, that error would arise after PosixEnter() had
registered the task but before the driver was uninitialized, allowing
the error to be saved in the registered task slot.  Now, if that error
occurs, PosixReturn() will invoke red_errnoptr() which will register
the task; if that is successful, the task slot will be used to store the
error (and if it's not successful, the global errno will be used).