posix: FildesOpen() error documentation minor fix
The documentation said that -RED_ENOENT would be returned if "the
@p pszPath argument, after removing the volume prefix, points to an
empty string."  However, the "after removing the volume prefix" clause
of that statement is incorrect.  If pszPath points to a string which
contains only the volume prefix, then that is taken to be a reference
to the root directory of that volume, rather than as an empty string
which should elicit an ENOENT error.