core: fail dirent deletions beyond EOF
This change causes RedDirEntryDelete() to return -RED_EINVAL when
ulDeleteIdx is beyond the end of the directory.  Previously, it would
write an empty directory entry in this scenario, which would expand the
directory with empty contents.  Callers of RedDirEntryDelete() would
prevent this from happening under normal circumstances, so this issue
would not likely occur in the real world.