posix: fix search permission check for . and ..
If permissions are enabled, a "." or ".." path element should provoke
an EACCES error if the parent directory is non-searchable.

For normal names, PathWalkFollow() will call into RedCoreLookup(),
which will check whether the parent directory is searchable and return
an EACCES error if it's not.  However, PathWalkFollow() treats "." and
".."  as special cases, and the directory permissions were not being
checked on those code paths.  As a result, all path-based APIs which
only entered into a non-searchable directory via "." or ".." could
erroneously return success rather than the required EACCES error.

InodeMustBeDir() was already used to check whether the inode for the
path component prior to "." or ".." was a directory.  Modify and rename
that function so that it checks whether the inode is a searchable
directory, thereby fixing this issue.