imgbld: update stale comment about permissions
The comment "if the POSIX-like API ever begins to enforce permissions"
implies that the POSIX-like API does not enforce permissions.  At the
time that comment was written, permissions had only been partially
implemented: they were stored but not enforced.  In the finished
implementation that we have today, permissions are enforced (if they
are enabled in the configuration).