core: use __FILE_NAME__ if it's defined
The CRITICAL_ERROR() macro takes __FILE__ as a parameter, which (with
some compilers) expands to a full path.  However, RedVolCriticalError()
only wants the basename (i.e., the final path component), and if
provided with a full path it will skip over the directory prefix
portion of the path to derive the basename.  But although the directory
prefix won't be used, the __FILE__ macro expansions will still include
it, so it will still be embedded into the binary as string data.  In
other words, string data which won't ever be used is included in the
binary.

GCC and Clang both implement __FILE_NAME__ as an extension.[1][2]  If
available, we can use it instead of __FILE__, which eliminates the
superfluous directory prefixes from the embedded strings, thereby
decreasing binary size.  It also allows us to remove a small amount of
code from OutputCriticalError().

[1] https://gcc.gnu.org/onlinedocs/cpp/Common-Predefined-Macros.html
[2] https://clang.llvm.org/docs/LanguageExtensions.html#builtin-macros