core: mount error if volume exceeds maximum size
With block sizes less than 16KB, the maximum volume size is constrained
by the number of blocks that the imap can address.  However, nothing was
enforcing these maximum sizes.  We allowed a volume to be formatted and
mounted even if the volume was above the maximum size and hence the imap
was too small for all the volume's blocks.  This erroneous configuration
would seem to work until we tried to allocate a block beyond the end of
the imap, at which time we would attempt to read/write beyond the end of
the metaroot's imap node bitmap.

Fix this problem by checking whether the number of imap nodes needed for
the block count is greater than the number of bits in the metaroot imap
node bitmap.  If it is, return an error, causing a format/mount error:
the user will need to increase the block size or decrease the block
count.

Additionally, update the parameter checking in imapextern.c to avoid
reading/writing beyond the end of the metaroot bitmap in cases where the
ulImapNode index is out-of-range.  While ulImapNode shouldn't ever be
out of range now that this bug has been fixed, it's our policy to be
paranoid with parameter validation.

Fixes: https://github.com/tuxera/reliance-edge/discussions/41