core: fix NULL ptr dereference in freserve code
CountSparseBlocks(), a subfunction of the red_freserve() API, could
dereference a NULL pointer.  The dereferenced NULL pointer could be
either pInode->pDindir or pInode->pIndir: in this explanation, we'll
assume the latter, since that's what was observed in practice.

The scenario is that the first call to SeekInode() returns -RED_ENODATA
because the block offset is sparse.  That error is cleared so the
control flow continues into the next block.  There, we populate a number
of "uPrev*Entry" values, which are used in the subsequent loop in order
to count sparse blocks.  However, it's possible for pInode->uIndirEntry
to be populated with a valid index (_not_ COORD_ENTRY_INVALID) while
pInode->pIndir is NULL.  This can occur when we are within the range of
the file where an indirect block is necessary, but the entire range of
the file represented by that indirect block is sparse, so no block yet
exists for the indirect, and so no indirect exists to be be buffered.

The solution is simple: just check whether pInode->pIndir is NULL.  When
it's NULL, we already know that the data block is sparse, and the
uPrevIndirEntry value was already initialized for a sparse block.