mount: discard master block buffer
The master block buffer is never used again after the master block is
mounted, so discard it.  That makes the buffer into the LRU buffer,
improving buffer efficiency -- particularly if there are other mounted
volumes which already have data in the buffer module.

As a consequence, RedBufferDiscard() is no longer conditionally
compiled: it always exists.