Add INDIRS_EXIST and DINDIRS_EXIST macros
Whether indirect nodes and double indirect nodes exist with the
configured inode pointers affects whether a lot of code needs to exist,
and thus preprocessor conditions for the existence of those nodes are
used in many places.  In my opinion, the preprocessor conditions that
were used previously were less than self-explanatory, so this commit
creates macros specifically for those conditions and updates the
preprocessor conditioning to use them.