Remove references to FSF's Boston office
On August 31st, the Free Software Foundation closed their Boston
office.[1]  On September 18th, GPLv2 license was edited to replace
the street address of their former office with a fsf.org URL, and
the recommended notice text was edited to replace the street address
with a gnu.org URL.[2]  In addition to replacing the street address
with URLs, the example name of "Ty Coon" was changed to "Moe Ghoul".

Make the same changes to our copies of the GPLv2 license and notice
texts.  This commit only changes text files and comments -- no
functional changes.

[1] <https://www.fsf.org/blogs/community/fsf-office-closing-party>

[2] <https://www.gnu.org/licenses/old-licenses/gpl-2.0.html>
Here is the last version of the page with the old text:
<https://web.archive.org/web/20240918031648/https://www.gnu.org/licenses/old-licenses/gpl-2.0.html>
Here is the first version of the page with the new text:
<https://web.archive.org/web/20240918192615/https://www.gnu.org/licenses/old-licenses/gpl-2.0.html>