Use Dokany to support FUSE driver on Windows
Dokany [1] is a Windows library for userspace filesystem drivers,
similar to FUSE on POSIX operating systems.  Dokany has its own native
interface but it also has a wrapper allowing it to be used via a FUSE
interface.

Move fuse.c to a directory for cross-platform code and update it to
compile cleanly on Windows with Visual Studio.

Update the Windows host tools makefile to support building edge-u.exe
from fuse.c.  This code is conditional on Dokany building installed, as
indicated by the DokanLibrary2 environment variable.

[1] https://github.com/dokan-dev/dokany